Initial idea for boarding pass config format
This commit lays out the basic config format that I think I want
to build out for this feature flag library. Going with the
name `conductor`, I've chosen to call the feature flags `boarding
passes` and to be opinionated that those passes have a particular
timeframe where they are valid, after which they will expire and
cause exceptions to be thrown if they are called without explicitly
extending the expiration date.

This is done in order to prevent the boarding passes from getting
forgotten about and rotting away in the background.